Get rid of unsigned integer overflow in column calculation
While unsigned integer overflow is well-defined, Android sanitizers treat it
as an error. We also have one in the CHAR_HASH macro but that one won't be
easy to get rid of (apparently, turning that macro into an inline function
silences the sanitizer for some reason, at least in the current version).